Remediate Target Width KLV Element (#8)
Handle the case if the Target Width KLV element (tag 22) has a value length greater than two bytes.  Sometimes, systems will use four bytes for the target width.  If so, truncate the value to two bytes.
This may be an issue with other KLV elements and a more general solution may need to be implemented.